backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
PyJWT
pytesseract
Pillow
python-multipart